Update PostgreSQL version and group.
According to http://mvnrepository.com/artifact/postgresql/postgresql
the artifact has been moved from group "postgresql" to "org.postgresql".
Unfortunately using the old version in PostGIS JDBC and declaring a
newer version in a project using it will result in the inclusion of two
PostgreSQL JDBC with different versions in the resulting application.

I had to change the acceptsURL functions of the DriverWrappers as well
to not declare throwing an SQLException as
org.postgresql.Driver.acceptsURL does no more throw it.